move api to ai2_internal
