+++ ai2_internal/bibentry_detection_predictor/interface.py
@@ -10,8 +10,9 @@ from typing import List
 
 from pydantic import BaseModel, BaseSettings, Field
 
+from ai2_internal import api
 from mmda.predictors.d2_predictors.bibentry_detection_predictor import BibEntryDetectionPredictor
-from mmda.types import api, image
+from mmda.types import image
 from mmda.types.document import Document
